worker: Move RowAggregator into its own file.
This commit only moves code and doesn't change it.
